May 07, 2020 | 12:16 PM
Made a popup window appear when clicking any of the folders.

Tried to add a recycler view to the popup window but this failed as when
the popup opens, the fragment is detached from the main activity and
thus, the context cannot be returned.

Changed directory structure of java classes.