May 19, 2020 | 06:58 PM
Added a recycler view for displaying boards (current preview is a
placeholder text).

Added a new activity that displays the name of the board it is triggered
from.

Added the creation of a directory to store data about the board in a
text file. May change the format of this file later.

Added a custom toolbar to the main activity.

Fixed the space above on of the recycler views.