May 24, 2020 | 08:04 PM
Scrapped the canvas method for displaying a board and switched it for a
system that adds a new image view to an intermediary relative view in
which it can then be moved.

Added a saving and loading system (note is currently loaded at 100, 100
for debugging purposes).

Added a new note picker (not yet active but there) in which the user can
select a new note to add. First they select a folder and then the new
note.

Added a system to prevent the user from dragging a note out of bounds.
It's not perfect and it looks kind of glitchy but it works.

Changed some colour values to attempt at fixing a bug in the light
theme. Didn't work.