Jun 13, 2020 | 03:36 PM
Added a system to create new boards.

Added a context menu (long press) system to boards, folders, and notes
which allows them to be manipulated (moved, deleted, renamed, etc.).
This renaming needs the names of the items to be replaced in save files
as well.

Added two new recycler views for the folder and note picker activities
instead of recycling (kek) old ones.

Added lots of new documentation.