Jun 14, 2020 | 11:34 PM
Fixed the recrop and cancel buttons.

Changed the theming of the buttons.

Delete reference of a note in a board when deleting the note.

Add rotating buttons to rotate the note after cropping it.

Added naming filter to stop it people from breaking the app with commas
and slashes.

Added note sharing.

Changed the name of the file provider to match the package name.

Changed default screen to note organizer.

Changed the colour of the card views in the recycler views.

Made it so that it always exits to the right screen in the main
activity.

Fixed exiting from move activity.